backend/requirements.txt
fastapi
uvicorn
httpx
yfinance
python-dotenv
openai
